--- scripts/fresh_install.sh
@@ -1,36 +1,0 @@
-#!/bin/bash
-
-username=$1
-password=$2
-
-if [ $(which jenkins) ]; then
-	echo "Jenkins already exists on this device."
-	exit 1
-fi
-
-# Install java
-apt update
-apt install -y curl fontconfig openjdk-17-jre
-
-# Install jenkins
-wget -O /usr/share/keyrings/jenkins-keyring.asc \
-  https://pkg.jenkins.io/debian-stable/jenkins.io-2023.key
-echo deb [signed-by=/usr/share/keyrings/jenkins-keyring.asc] \
-  https://pkg.jenkins.io/debian-stable binary/ | tee \
-  /etc/apt/sources.list.d/jenkins.list > /dev/null
-apt update
-apt install jenkins
-
-# Enable service
-systemctl daemon-reload
-systemctl enable jenkins
-
-# Start jenkins
-systemctl start jenkins &
-
-# Auth for the first time
-wget http://localhost:8080/jnlpJars/jenkins-cli.jar -P "../artifacts"
-pat=$(cat /var/lib/jenkins/secrets/initialAdminPassword)
-#java -jar jenkins-cli.jar -s http://192.168.1.219:8080 -auth admin:$pat
-
-echo "jenkins.model.Jenkins.instance.securityRealm.createAccount('\""$username"\"', '\""$password"\"')" | java -jar jenkins-cli.jar -auth admin:$pat -s http://localhost:8080/ groovy =
